Refactor <Motion> using mocks
